fix: Use append for SSH keys (audit requirement)
